+++ assets/styles/molecules/Desktop/Header/DropDown.module.scss
@@ -14,14 +14,3 @@
   transform: translateX(10px);
   opacity: 1;
 }
-
-.darkmode_container {
-  @extend .w_100, .h_100, .flex_center_center;
-  gap: 40px;
-  cursor: pointer;
-}
-
-.darkmode_text {
-  font-size: var(--font-size-8);
-  font-weight: var(--font-weight-semibold);
-}
