+++ assets/styles/molecules/Desktop/Header/Nav/Darkmode.module.scss
@@ -1,0 +1,12 @@
+@import "../../../../bass.scss";
+
+.darkmode_container {
+  @extend .w_100, .h_100, .flex_center_center;
+  gap: 40px;
+  cursor: pointer;
+}
+
+.darkmode_text {
+  font-size: var(--font-size-8);
+  font-weight: var(--font-weight-semibold);
+}
